youtube-chat-backend/requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
langchain
langchain-openai
langchain-community
faiss-cpu
openai
pydantic